+++ noodles-util/src/alignment.rs
@@ -1,8 +1,12 @@
 //! I/O for alignment formats.
 
+mod compression_method;
 mod format;
 pub mod indexed_reader;
 pub mod reader;
 pub mod writer;
 
-pub use self::{format::Format, indexed_reader::IndexedReader, reader::Reader, writer::Writer};
+pub use self::{
+    compression_method::CompressionMethod, format::Format, indexed_reader::IndexedReader,
+    reader::Reader, writer::Writer,
+};