+++ noodles-util/src/alignment/compression_method.rs
@@ -1,0 +1,6 @@
+/// A compression method.
+#[derive(Clone, Copy, Debug, Eq, PartialEq)]
+pub enum CompressionMethod {
+    /// BGZF.
+    Bgzf,
+}