+++ noodles-util/src/alignment/writer/builder.rs
@@ -14,17 +14,24 @@ use noodles_fasta as fasta;
 use noodles_sam as sam;
 
 use super::Writer;
-use crate::alignment::Format;
+use crate::alignment::{CompressionMethod, Format};
 
 /// An alignment writer builder.
 #[derive(Default)]
 pub struct Builder {
+    compression_method: Option<Option<CompressionMethod>>,
     format: Option<Format>,
     reference_sequence_repository: fasta::Repository,
     block_content_encoder_map: BlockContentEncoderMap,
 }
 
 impl Builder {
+    /// Sets the compression method.
+    pub fn set_compression_method(mut self, compression_method: Option<CompressionMethod>) -> Self {
+        self.compression_method = Some(compression_method);
+        self
+    }
+
     /// Sets the format of the output.
     ///
     /// # Examples
@@ -104,7 +111,12 @@ impl Builder {
         let src = src.as_ref();
 
         if self.format.is_none() {
-            self.format = detect_format_from_path_extension(src);
+            if let Some((format, compression_method)) =
+                detect_format_and_compression_method_from_path_extension(src)
+            {
+                self.format = Some(format);
+                self.compression_method = Some(compression_method);
+            }
         }
 
         let file = File::create(src).map(BufWriter::new)?;
@@ -129,36 +141,49 @@ impl Builder {
     {
         let format = self.format.unwrap_or(Format::Sam);
 
-        let inner: Box<dyn sam::AlignmentWriter> = match format {
-            Format::Sam => Box::new(sam::Writer::new(writer)),
-            Format::SamGz => Box::new(sam::Writer::new(bgzf::Writer::new(writer))),
-            Format::Bam => Box::new(bam::Writer::new(writer)),
-            Format::Cram => Box::new(
+        let compression_method = match self.compression_method {
+            Some(compression_method) => compression_method,
+            None => match format {
+                Format::Sam | Format::Cram => None,
+                Format::Bam => Some(CompressionMethod::Bgzf),
+            },
+        };
+
+        let inner: Box<dyn sam::AlignmentWriter> = match (format, compression_method) {
+            (Format::Sam, None) => Box::new(sam::Writer::new(writer)),
+            (Format::Sam, Some(CompressionMethod::Bgzf)) => {
+                Box::new(sam::Writer::new(bgzf::Writer::new(writer)))
+            }
+            (Format::Bam, Some(CompressionMethod::Bgzf)) => Box::new(bam::Writer::new(writer)),
+            (Format::Cram, None) => Box::new(
                 cram::writer::Builder::default()
                     .set_reference_sequence_repository(self.reference_sequence_repository)
                     .set_block_content_encoder_map(self.block_content_encoder_map)
                     .build_with_writer(writer),
             ),
+            (_, _) => todo!(),
         };
 
         Writer { inner }
     }
 }
 
-fn detect_format_from_path_extension<P>(path: P) -> Option<Format>
+fn detect_format_and_compression_method_from_path_extension<P>(
+    path: P,
+) -> Option<(Format, Option<CompressionMethod>)>
 where
     P: AsRef<Path>,
 {
     let ext = path.as_ref().file_name().and_then(|ext| ext.to_str())?;
 
     if ext.ends_with("sam") {
-        Some(Format::Sam)
+        Some((Format::Sam, None))
     } else if ext.ends_with("sam.gz") || ext.ends_with("sam.bgz") {
-        Some(Format::SamGz)
+        Some((Format::Sam, Some(CompressionMethod::Bgzf)))
     } else if ext.ends_with("bam") {
-        Some(Format::Bam)
+        Some((Format::Bam, Some(CompressionMethod::Bgzf)))
     } else if ext.ends_with("cram") {
-        Some(Format::Cram)
+        Some((Format::Cram, None))
     } else {
         None
     }
@@ -169,24 +194,24 @@ mod tests {
     use super::*;
 
     #[test]
-    fn test_detect_format_from_path_extension() {
+    fn test_detect_format_and_compression_method_from_path_extension() {
         assert_eq!(
-            detect_format_from_path_extension("out.sam"),
-            Some(Format::Sam)
+            detect_format_and_compression_method_from_path_extension("out.sam"),
+            Some((Format::Sam, None))
         );
         assert_eq!(
-            detect_format_from_path_extension("out.sam.gz"),
-            Some(Format::SamGz)
+            detect_format_and_compression_method_from_path_extension("out.sam.gz"),
+            Some((Format::Sam, Some(CompressionMethod::Bgzf)))
         );
         assert_eq!(
-            detect_format_from_path_extension("out.bam"),
-            Some(Format::Bam)
+            detect_format_and_compression_method_from_path_extension("out.bam"),
+            Some((Format::Bam, Some(CompressionMethod::Bgzf)))
         );
         assert_eq!(
-            detect_format_from_path_extension("out.cram"),
-            Some(Format::Cram)
+            detect_format_and_compression_method_from_path_extension("out.cram"),
+            Some((Format::Cram, None))
         );
 
-        assert!(detect_format_from_path_extension("out.fa").is_none());
+        assert!(detect_format_and_compression_method_from_path_extension("out.fa").is_none());
     }
 }